- add fixes for linux build [TESTS] - checked if wb builds without problems - checked if wb is starting - checked if code completion and error reporting in sql editor works

+++ backend/wbpublic/sqlide/sql_editor_be.cpp
@@ -696,7 +696,7 @@ void Sql_editor::char_added(int char_code)
 
 //--------------------------------------------------------------------------------------------------
 
-void Sql_editor::dwell_event(bool started, int position, int x, int y)
+void Sql_editor::dwell_event(bool started, size_t position, int x, int y)
 {
   if (started)
   {
