+++ backend/wbpublic/sqlide/sql_editor_be.h
@@ -268,7 +268,7 @@ private:
 
   void text_changed(int position, int length, int lines_changed, bool added);
   void char_added(int char_code);
-  void dwell_event(bool started, int position, int x, int y);
+  void dwell_event(bool started, size_t position, int x, int y);
 
   void setup_editor_menu();
   void editor_menu_opening();
